% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berrylogo.r
\name{berrylogo.PFM}
\alias{berrylogo.PFM}
\title{Generates a Berry logo in ggplot from a TFBSTools
Positional Frequency Matrix, which is count based}
\usage{
berrylogo.PFM(pfm, ...)
}
\arguments{
\item{pfm}{matrix of fractional nucleotide counts
here rows are A, C, G, T and columns are position}

\item{...}{additional arguments}
}
\description{
Generates a Berry logo in ggplot from a TFBSTools
Positional Frequency Matrix, which is count based
}
