% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berrylogo.r
\name{berrylogo.PWM}
\alias{berrylogo.PWM}
\title{Generates a Berry logo in ggplot TFBSTools
Positional Weight Matrix, which is log2-based}
\usage{
berrylogo.PWM(pwm)
}
\arguments{
\item{pwm}{matrix of fractional nucleotide frequencies in log2 scale
here rows are A, C, G, T and columns are position}
}
\value{
A ggplot2 object, use print() to display
}
\description{
Generates a Berry logo in ggplot TFBSTools
Positional Weight Matrix, which is log2-based
}
