% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berrylogo.r
\name{berrylogo}
\alias{berrylogo}
\title{Get the total width of an ascii ruler}
\usage{
berrylogo(x, ...)
}
\arguments{
\item{x}{the asciiruler}

\item{...}{additional arguments}
}
\description{
Get the total width of an ascii ruler
}
