% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berrylogo.r
\name{berrylogo.PFM}
\alias{berrylogo.PFM}
\title{Generates a Berry logo in ggplot from a TFBSTools
Positional Frequency Matrix, which is count based}
\usage{
berrylogo.PFM(pwm, ...)
}
\arguments{
\item{pwm}{matrix of fractional nucleotide counts
here rows are A, C, G, T and columns are position}

\item{gc_content}{optional the GC content of the source genome}

\item{zero}{optional a placeholder estimate for true frequency when 0 is observed}

\item{base}{the log base to use}
}
\description{
Generates a Berry logo in ggplot from a TFBSTools
Positional Frequency Matrix, which is count based
}
