% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berrylogo.r
\name{berrylogo.PWM}
\alias{berrylogo.PWM}
\title{Generates a Berry logo in ggplot TFBSTools
Positional Weight Matrix, which is log2-based}
\usage{
berrylogo.PWM(pwm)
}
\arguments{
\item{pwm}{matrix of fractional nucleotide frequencies in log2 scale
here rows are A, C, G, T and columns are position}
}
\value{
A ggplot2 object, use print() to display
}
\description{
Generates a Berry logo in ggplot TFBSTools
Positional Weight Matrix, which is log2-based
}
\examples{
BiocInstaller::biocLite("TFBSTools",suppressUpdates=TRUE,suppressAutoUpdate=TRUE,ask=FALSE)
pfm <- TFBSTools::PFMatrix(ID="MA0004.1", name="Arnt", matrixClass="Zipper-Type", 
  strand="+", bg=c(A=0.25, C=0.25, G=0.25, T=0.25), 
  tags=list(family="Helix-Loop-Helix", species="10090", 
    tax_group="vertebrates",
    medline="7592839", type="SELEX", ACC="P53762", 
    pazar_tf_id="TF0000003",
    TFBSshape_ID="11", TFencyclopedia_ID="580"),
    profileMatrix=matrix(c(4L,  19L, 0L,  0L,  0L,  0L,
                        16L, 0L,  20L, 0L,  0L,  0L,
                        0L,  1L,  0L,  20L, 0L,  20L,
                        0L,  0L,  0L,  0L,  20L, 0L), 
                     byrow=TRUE, nrow=4, 
                     dimnames=list(c("A", "C", "G", "T")))
)
pwm <- TFBSTools::toPWM(pfm, type="log2probratio", pseudocounts=0.8)
}
